create retro discussion thread header
